outreach/test_client/requirements.txt
fastapi
uvicorn[standard]==0.34.0
jinja2==3.1.2
python-jose[cryptography]==3.3.0

deprecated>=1.2.14
a2a-sdk==0.2.5
# HTTP client for A2A communication
httpx==0.28.1